refactor(core): move mode parsing contracts to shared package
Why:
- operationPolicy と readIndexMode の文字列解釈がCLI側に残っており、契約正本が分散していたため。

What Changed:
- Ucli.Contracts.Configuration に OperationPolicy / ReadIndexMode enum と各 Codec を追加。
- UcliConfigStore と ReadIndexModeResolver の手書きパースを Contracts codec 参照へ置換。
- CLI側の重複 enum 定義を削除し、参照先を Contracts enum に統一。

Impact:
- mode文字列契約の単一正本が Contracts に揃い、CLI/Unity 間の解釈ずれリスクを低減。
