+++ src/Ucli/Operations/OperationAuthorizationService.cs
@@ -1,5 +1,6 @@
 using System.Text.RegularExpressions;
 using MackySoft.Ucli.Configuration;
+using MackySoft.Ucli.Contracts.Configuration;
 
 namespace MackySoft.Ucli.Operations;
